refactor local build service
